claim a broader expertise for RSEs
the original text here only mentioned "software development", which I
changed to "software engineering", because it includes more than "just
the development". This is not quite in line with the given reference, as
this says "exceptional individuals with combined expertise in programming and a
solid knowledge of the research environmen", but I would claim we should
rather state is this way than to use the term "programming". RSE is much
more than that.