Do not overemphesize 'technical'
This is not needed in those sentences, and it focuses too much on those
RSEs exclusively that choose this as their main or even sole work,
without every doing research, not even as part of a research team. It
kind of puts them on a similar level as an electrician or computer
network specialist: research also couldn't happen without those, they
are established professions, but they are a lot less integrated into
research.